+++ Changelog.md
@@ -25,8 +25,9 @@ Bug Fixes:
 * When defining `let` methods that overwrite an existing method, prevent
   a warning being issued by removing the old definition. (Jon Rowe, #2593)
 * Prevent warning on Ruby 2.6.0-rc1 (Keiji Yoshimi, #2582)
-* Fix `config.define_derived_metadata` so that it supports cascades.
-  (Myron Marston, #2630).
+* Fix `config.define_derived_metadata` so that it supports cascades and
+  is not triggered until metadata has been assigned to the example or
+  example group (Myron Marston, #2630, #2635).
 
 ### 3.8.0 / 2018-08-04
 [Full Changelog](http://github.com/rspec/rspec-core/compare/v3.7.1...v3.8.0)